Added more higher-level functions, minor fixes/improvements
Added a bunch more higher-level get/set functions. These need to be tested. Also re-ordered these functions in the headers to better match their order in terms of the address layout.

Address #defines are now uint8 type. This was required for the 0 address not to be ambiguous, so the rest were converted too. Might save a small amount of memory too depending on what the compiler does.

Signed-off-by: horchler <adh9@case.edu>